refactor(data-thesesul): refactor code
