+++ services/data-thesesul/v1/thesesul/thesesul_functions.py
@@ -1,0 +1,124 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import os
+import subprocess
+import requests
+import sys
+import json
+import re
+import zipfile
+
+# Chemin du répertoire temporaire
+TMP_DIR = '/tmp'
+
+retrieve_id = str(sys.argv[sys.argv.index('-p') + 1])
+OUTPUT_DIR = os.path.join(TMP_DIR, "retrieve", retrieve_id, "datas")
+os.makedirs(OUTPUT_DIR, exist_ok=True)
+
+
+def write_error_in_logs(log_file, sudoc_id, error):
+    with open(log_file, 'a') as f_log:
+        f_log.write(f"{sudoc_id},{error}"+ '\n')
+
+      
+def delete_files():
+    to_dels = [f"sudoc_file-{retrieve_id}.xml", f"conditor_file-{retrieve_id}.xml", f"tei_hal-{retrieve_id}.xml"]
+    for to_del in to_dels:
+        to_del = os.path.join(TMP_DIR, to_del)
+        if os.path.exists(to_del):
+            os.remove(to_del)
+
+
+def get_xml_from_sudoc(sudoc_id):
+    url = f"https://www.sudoc.fr/{sudoc_id}.xml"
+    output_path = os.path.join(TMP_DIR, f"sudoc_file-{retrieve_id}.xml")
+    
+    response = requests.get(url)
+    
+    if response.status_code != 200:
+        return None
+    else: 
+        with open(output_path, 'wb') as file:
+            file.write(response.content)
+        return output_path
+                
+
+def download_pdf(res_file):
+    with open(res_file, "r") as f:
+        file = f.read()
+        
+    pdf_pattern = re.compile(r'target=\".*?\.pdf\"')
+    matches = pdf_pattern.findall(file)
+    
+    if not any(matches):
+        return None
+    
+    output_pdf = res_file.replace("xml", "pdf")
+    url = "http://docnum.univ-lorraine.fr/public/" + matches[0].split('"')[1]
+    try:
+        response = requests.get(url)
+        if response.status_code!=200:
+            return None
+            
+        else:
+            with open(output_pdf, 'wb') as f:
+                f.write(response.content)
+            return output_pdf
+        
+    except requests.exceptions.RequestException as e:
+        return None
+
+
+def transform_sudoc_to_conditor(sudoc_file):
+    conditor_file = os.path.join(TMP_DIR, f"conditor_file-{retrieve_id}.xml")
+    cmd = ['xsltproc', '-o', conditor_file, 'v1/xsl-files/SudocThese2Conditor.xsl', sudoc_file]
+    result = subprocess.run(cmd, capture_output=True)
+    if result.returncode != 0:
+        return None
+    return conditor_file
+
+
+def transform_conditor_to_tei_hal(conditor_file):
+    tei_hal_file = os.path.join(TMP_DIR, f"tei_hal-{retrieve_id}.xml")
+    cmd = ['xsltproc', '-o', tei_hal_file, 'v1/xsl-files/LORR_TheseExe2HAL.xsl', conditor_file]
+    result = subprocess.run(cmd, capture_output=True)
+    if result.returncode != 0:
+        return None
+    return tei_hal_file
+
+
+def validate_tei_hal(tei_hal_file):
+    cmd = f'xmllint --noout {tei_hal_file} --schema v1/xsl-files/aofr3.xsd'
+    result = subprocess.run(cmd, shell=True, capture_output=True)
+
+    if result.returncode != 0:
+        return False
+    return True
+
+
+def modify_target_in_xml(file_path, sudoc_id):
+    with open(file_path, 'r', encoding='utf-8') as file:
+        content = file.read()
+
+    # use a regular expression to change a field
+    modified_content = re.sub(r'target=".*?\.pdf"', f'target="{sudoc_id}.pdf"', content)
+
+    with open(file_path, 'w', encoding='utf-8') as file:
+        file.write(modified_content)
+
+
+def create_zip(sudoc_id):
+    files_to_zip = [os.path.join(OUTPUT_DIR,f"{sudoc_id}.xml"), os.path.join(OUTPUT_DIR,f"{sudoc_id}.pdf")]
+    output_zip = os.path.join(OUTPUT_DIR,f'{sudoc_id}.zip')
+
+    try:
+        with zipfile.ZipFile(output_zip, 'w') as archive:
+            for fichier in files_to_zip:
+                archive.write(fichier, arcname=fichier.split('/')[-1])
+    except Exception as e:
+        sys.stderr.write(f"Failed to zip xml + pdf: {e}\n")
+        
+    if os.path.exists(os.path.join(OUTPUT_DIR,f"{sudoc_id}.pdf")):
+        os.remove(os.path.join(OUTPUT_DIR,f"{sudoc_id}.pdf"))
+        os.remove(os.path.join(OUTPUT_DIR,f"{sudoc_id}.xml"))
